v2.0: First major revision of the "Official" DOOM FAQ. Id gets involved, giving new information. The FAQ is rearranged. The FAQ is renamed from "The DOOM FAQ" to the "The 'Official' DOOM FAQ."
--

Version 2.0 of the FAQ adds a lot of updated information. This is because FAQ
v1.0 was only created from the DOOM press release and Usenet "hearsay." After
v1.0 was released, id contacted me, and offered to help with new releases.
So, here it is. This may or may not be the last "Official" DOOM FAQ before
the game is released, I'm not sure. I guess it just depends on how quickly
new information arrives.