v5.0: A major revision of the "Official" DOOM FAQ. DOOM is released. FAQ is completely rewritten. FAQ tells about troubleshooting, cheating, and add-on software. New additions to DOOM iNsAnItY.
--

FINALLY! After 344 days of waiting after the press release, DOOM was released
on time. The game we all have been waiting for is BETTER than anyone even
expected it to be! I would like to thank all the "Usenetters" out there for not
making me eat my (or other people's) socks, and for not nuking my house. I know
there was a temptation, but id Software saved me!  Anyway, I hope you enjoy
this version of the FAQ. I hope this one answers all the questions you can
possibly dream up about DOOM! After all, it almost doubles the size of v2.6 of
the FAQ, and is about one-twentieth of the size of the actual game!

Currently, I have only seen the shareware version of the game. A small amount
of the information may not pertain to the mail-order or commercial version, but
most is accurate for all versions.