v5.5: A standard revision of the "Official" DOOM FAQ.  DOOM v1.2 is released, and information on modem play is added.  DOOM iNsAnItY is released seperately. A complete list of add-on software, modem init strings, and troubleshooting guidelines are added.
--

It has been more than two months since the last release of the FAQ, and there
is a lot of good reasons.  I promised to have a FAQ out at around the same
time that DOOM v1.2 was released, but DOOM v1.2 was delayed for about a month
and a half.  I had a hard drive crash, and all of the work I had previously
done on v5.5 of the FAQ was gone.  My E-mail box keeps overflowing, and at the
rate the information is coming in, there would not be ANY way for me to
release the FAQ if I had not drawn the line somewhere.  Anyway, here it is,
the "Official" DOOM FAQ v5.5.

DOOM iNsAnItY was removed (sorry) due to size complaints, but the last edition
of DOOM iNsAnItY has been released with this FAQ.  Look for it in the file
"dinsan55.txt" on the Internet or "dinsan55.zip" on BBSs.  DOOM v1.2 has been
released, which has caused one huge problem: modem support.  After weeding
through hundreds of messages, and with the help of a few others, a list of
modem strings that work with DOOM is included in this FAQ.  I hope one of them
works for you!  "Keep on DOOMing!"